refactor : utils 패키지 변경
