+++ src/main/java/techcourse/fakebook/service/utils/UserAssembler.java
@@ -1,4 +1,4 @@
-package techcourse.fakebook.utils;
+package techcourse.fakebook.service.utils;
 
 import techcourse.fakebook.domain.user.User;
 import techcourse.fakebook.service.dto.UserDto;
